vmwatcher: rename response.go to checker, other minor format fixes
